Move isTrivialWatchlist out of "use server" module
Server Action files require all exports to be async functions.
isTrivialWatchlist is a sync helper, which broke the production
build after #2177. Move it to a plain module.
