+++ apps/web/src/lib/actions/watchlists.ts
@@ -21,6 +21,7 @@ import {
   deleteWatchlist as tsDeleteWatchlist,
   updateWatchlistField as tsUpdateWatchlistField,
 } from "@/lib/search/typesense-watchlist";
+import { isTrivialWatchlist } from "@/lib/watchlist-utils";
 
 // ── Types ───────────────────────────────────────────────────────────
 
@@ -88,29 +89,6 @@ export type WatchlistPostingEntry = {
     icon: string | null;
   };
 };
-
-// A watchlist is "trivial" when it carries no meaningful filters and tracks
-// no companies — effectively a blank shell. We exclude these from public
-// listings and search engines so they don't dilute the index.
-// `anyCompany` and `salaryCurrency` alone don't count (they're defaults/prefs).
-export function isTrivialWatchlist(
-  filters: WatchlistFilters | null | undefined,
-  companyCount: number,
-): boolean {
-  if (companyCount > 0) return false;
-  const f = filters ?? {};
-  return !(
-    f.keywords?.length ||
-    f.locationSlugs?.length ||
-    f.occupationSlugs?.length ||
-    f.senioritySlugs?.length ||
-    f.technologySlugs?.length ||
-    f.salaryMin != null ||
-    f.salaryMax != null ||
-    f.experienceMin != null ||
-    f.experienceMax != null
-  );
-}
 
 // ── Actions ─────────────────────────────────────────────────────────
